+++ src/switchable/src/circular.js
@@ -10,7 +10,6 @@ KISSY.add('switchable/circular', function (S, DOM, Anim, Switchable) {
         TOP = 'top',
         EMPTY = '',
         PX = 'px',
-        FORWARD = 'forward',
         BACKWARD = 'backward',
         SCROLLX = 'scrollx',
         SCROLLY = 'scrolly';
@@ -25,37 +24,37 @@ KISSY.add('switchable/circular', function (S, DOM, Anim, Switchable) {
     /**
      * 循环滚动效果函数
      */
-    function circularScroll(fromEls, toEls, callback, index, direction) {
+    function circularScroll(fromEls, toEls, callback,fromIndex, index, direction) {
         var self = this,
             cfg = self.config,
             len = self.length,
-            // 2012-03-08 yiminghe@gmail.com
-            // 当前 panel 确切在的位置，而不是
-            // panel 将要的位置 activeIndex (trigger 当前位置)
-            completedIndex = self.completedIndex,
             isX = cfg.scrollType === SCROLLX,
             prop = isX ? LEFT : TOP,
             viewDiff = self.viewSize[isX ? 0 : 1],
             diff = -viewDiff * index,
+            panels = self.panels,
             props = {},
             isCritical,
             isBackward = direction === BACKWARD;
+
         // 从第一个反向滚动到最后一个 or 从最后一个正向滚动到第一个
-        isCritical = (isBackward && completedIndex === 0 && index === len - 1)
-            || (direction === FORWARD && completedIndex === len - 1 && index === 0);
+        isCritical = (isBackward && fromIndex === 0 && index === len - 1)
+            || (!isBackward && fromIndex === len - 1 && index === 0);
+
+        // 开始动画
+        if (self.anim) {
+            self.anim.stop();
+            // 清除 relative 定位的单项，防止速度过快有未清除的情况
+            resetPosition.call(self, panels, 1, prop, viewDiff);
+            resetPosition.call(self, panels, 0, prop, viewDiff);
+        }
 
         if (isCritical) {
             // 调整位置并获取 diff
-            diff = adjustPosition.call(self, self.panels, index, isBackward, prop, viewDiff);
+            diff = adjustPosition.call(self, panels, isBackward, prop, viewDiff);
         }
 
         props[prop] = diff + PX;
-
-        // 开始动画
-
-        if (self.anim) {
-            self.anim.stop();
-        }
 
         if (fromEls) {
             self.anim = new Anim(self.content,
@@ -65,7 +64,7 @@ KISSY.add('switchable/circular', function (S, DOM, Anim, Switchable) {
                 function () {
                     if (isCritical) {
                         // 复原位置
-                        resetPosition.call(self, self.panels, index, isBackward, prop, viewDiff);
+                        resetPosition.call(self, panels, isBackward, prop, viewDiff, 1);
                     }
                     // free
                     self.anim = undefined;
@@ -82,14 +81,13 @@ KISSY.add('switchable/circular', function (S, DOM, Anim, Switchable) {
     /**
      * 调整位置
      */
-    function adjustPosition(panels, index, isBackward, prop, viewDiff) {
+    function adjustPosition(panels, isBackward, prop, viewDiff) {
         var self = this, cfg = self.config,
             steps = cfg.steps,
             len = self.length,
             start = isBackward ? len - 1 : 0,
             from = start * steps,
-            to = (start + 1) * steps,
-            i;
+            to = (start + 1) * steps;
 
         // 调整 panels 到下一个视图中
         var actionPanels = panels.slice(from, to);
@@ -103,7 +101,7 @@ KISSY.add('switchable/circular', function (S, DOM, Anim, Switchable) {
     /**
      * 复原位置
      */
-    function resetPosition(panels, index, isBackward, prop, viewDiff) {
+    function resetPosition(panels, isBackward, prop, viewDiff, setContent) {
         var self = this,
             cfg = self.config,
             steps = cfg.steps,
@@ -117,9 +115,15 @@ KISSY.add('switchable/circular', function (S, DOM, Anim, Switchable) {
         DOM.css(actionPanels, POSITION, EMPTY);
         DOM.css(actionPanels, prop, EMPTY);
 
-        // 瞬移到正常位置
-        DOM.css(self.content, prop, isBackward ? -viewDiff * (len - 1) : EMPTY);
+        if (setContent) {
+            // 瞬移到正常位置
+            DOM.css(self.content, prop, isBackward ? -viewDiff * (len - 1) : EMPTY);
+        }
     }
+
+    Switchable.adjustPosition = adjustPosition;
+
+    Switchable.resetPosition = resetPosition;
 
     /**
      * 添加插件
@@ -146,6 +150,9 @@ KISSY.add('switchable/circular', function (S, DOM, Anim, Switchable) {
 }, { requires:["dom", "anim", "./base", "./effect"]});
 
 /**
+ * 2012-04-12 yiminghe@gmail.com
+ *  - 修复速度过快从 0 到最后或从最后到 0 时的小bug
+ *
  * 承玉：2011.06.02 review switchable
  *
  * TODO: